fix: route managed model scope through Execute request
